# HDDS-10311. Speed up TestOmMetrics