@will-sh 2024-03-20
Thx for your advice! @adoroszlai 
Now each test time looks better and cluster only start/stop once:
[INFO] Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 40.724 s - in org.apache.hadoop.ozone.om.TestOmMetrics
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testDirectoryOps(BucketLayout)[1]  Time elapsed: 4.781 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testDirectoryOps(BucketLayout)[2]  Time elapsed: 0.38 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testBucketOps  Time elapsed: 0.82 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testVolumeOps  Time elapsed: 0.192 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testSnapshotOps  Time elapsed: 3.365 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testAclOperationsHA  Time elapsed: 0.096 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testAclOperations  Time elapsed: 0.028 s
[INFO] org.apache.hadoop.ozone.om.TestOmMetrics.testKeyOps  Time elapsed: 0.494 s